+++ ton_sdk_cli/tests/cli.rs
@@ -1,0 +1,8 @@
+use assert_cmd::prelude::*;
+use predicates::prelude::*;
+use std::process::Command;
+
+#[test]
+fn test_config() {
+
+}